// src/services/-workflows-{key}-{entityId}-available-transitions-get.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Shopping Platform API
 * Production-grade e-commerce backend
 * OpenAPI spec version: 1.0
 */
export interface UtilsResponse {
  code?: number;
  error?: string;
  message?: string;
  success?: boolean;
}

export interface DtoStateView {
  code?: string;
  color?: string;
  id?: number;
  is_final?: boolean;
  is_initial?: boolean;
  name?: string;
  sort_order?: number;
  text_color?: string;
}

export interface DtoTransitionView {
  event?: string;
  from_state?: DtoStateView;
  guard_key?: string;
  hook_key?: string;
  id?: number;
  is_active?: boolean;
  name?: string;
  required_role?: string;
  to_state?: DtoStateView;
}

export interface DtoAvailableTransitionsView {
  current_state?: DtoStateView;
  transitions?: DtoTransitionView[];
}

export type GetWorkflowsKeyEntityIdAvailableTransitions200 = UtilsResponse & {
  data?: DtoAvailableTransitionsView;
};